add median_of_two_sorted_arrays

+++ src/subtract_the_product_and_sum_of_digits_of_an_integer.rs
@@ -11,7 +11,7 @@ impl Solution {
             let x = n % 10;
             sum += x;
             product *= x;
-            n = n / 10 as i32;
+            n = n / 10;
         }
 
         product - sum
